requirements.txt
cycler==0.10.0
Cython==0.29.13
Django==2.2.4
django-env==0.2.1
djangorestframework==3.10.2
fredapi==0.4.2
hupper==1.8.1
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
pandas==0.25.1
PasteDeploy==2.0.1
patsy==0.5.1
plaster==1.0
plaster-pastedeploy==0.7
pyparsing==2.4.2
pyramid==1.10.4
python-dateutil==2.8.0
pytz==2019.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
sqlparse==0.3.0
statsmodels==0.10.1
timetools==0.6.0
translationstring==1.3
venusian==1.2.0
WebOb==1.8.5
zope.deprecation==4.4.0
zope.interface==4.6.0
